+++ CHANGELOG.md
@@ -6,6 +6,7 @@ General
 
 - Button components will now register gestures even if the background is clear
 - UIModels are re-structured and no longer depend on sub UI models
+- `ColorBook` is renamed to `VComponentsColorBook` to avoid issues with name-shadowing
 - `ImageBook` is no longer exposed to `public`, but UI models take icons instead
 
 VWrappedButton